grantreg contact-us page
(https://grantreg.com/contact-us page). It consist of a header, maiin content which include a fom to fill and a foote which contains links to contact the comoany and all one needs too kmow about the company. The header also contain two dropdowns on hover, one to know about the number of carts remaining and the other to know about the categories of products in the company